// src/app/vo/application.interface.ts
export interface Application {
  businessName: string;
  amountsToBorrow: string;
  creditCardVolume: string;
  monthlyDeposit: string;
  time: string;
  industry: string;

  street: string;
  city: string;
  state: string;
  zip: string;

  ficoScore: string;
  firstName: string;
  lastName: string;
  phone: string;
}